+++ app/views/playground.vue
@@ -20,8 +20,10 @@
 			</fieldset>
 			<fieldset>
 				<BoolStoreInput v-show="false" v-model="tokenizeStaff" sessionKey="lotus-tokenizeStaff" />
+				<BoolStoreInput v-show="false" v-model="rollVisible" sessionKey="lotus-rollVisible" />
 				<CheckButton content="&#x1f3b9;" v-model="tokenizeStaff" title="live staff" />
 				<CheckButton content="&#x1f3a8;" v-model="chromaticSymbols" title="chromatic symbols" />
+				<CheckButton content="&#x2633;" v-model="rollVisible" title="show MIDI roll" />
 				<button @click="togglePlayer" :disabled="!midiPlayer">{{midiPlayer && midiPlayer.isPlaying ? "&#x23f8;" : "&#x25b6;"}}</button>
 			</fieldset>
 		</header>
@@ -32,7 +34,7 @@
 			</div>
 			<div class="sheet-container" :class="{loading: engraving, dirty: engraverDirty, chromatic: chromaticSymbols}">
 				<SheetSimple v-if="svgDocuments && !tokenizeStaff" :documents="svgDocuments" />
-				<MidiRoll v-if="midiPlayer" :player="midiPlayer" :timeScale="4e-3" :height="120" />
+				<MidiRoll v-if="midiPlayer" v-show="rollVisible" :player="midiPlayer" :timeScale="4e-3" :height="120" />
 				<SheetLive v-if="tokenizeStaff && sheetContent"
 					:content="sheetContent"
 					:hashTable="svgHashTable"
@@ -96,6 +98,7 @@
 				midi: null,
 				chromaticSymbols: false,
 				midiPlayer: null,
+				rollVisible: false,
 			};
 		},