+++ inc/jsonRecovery.ts
@@ -5,7 +5,7 @@ const recoverJSON = (json: string | object, classDict) => {
 		json = JSON.stringify(json);
 
 	return JSON.parse(json, (_, value) => {
-		if (value.__prototype) {
+		if (value && (typeof value === "object") && value.__prototype) {
 			const Class = classDict[value.__prototype];
 			if (Class) {
 				const {__prototype, ...fields} = value;